// src/app/model/contact.ts
export interface Contact {
    id?: string;
    firstName?: string;
    middleName?: string;
    lastName?: string;
    jobTitle?: string;
    department?: string;
    companyName?: string;
    birthday?: string;
    email?: string;
    mobileNumber?: string;
    secondMobileNumber?: string;
    //address: string;
    street?: string;
    city?: string;
    state?: string;
    zipCode?: string;
    country?: string;
}